demonstrate a tragic unsoundness
It turns out that the current approach we take to builds - that is, to
changes like (\a -> new_elem:a) - is not actually sound. It seems to
work when the function receiving the change is of type [a] -> b, but not
if it is of type [a] -> [b].

So that sucks.