incrementalise map and filter by hand
When in doubt, cheat. My current automatic incrementalizer is unsound in
the face of recursive functions, so implement the buggers - map and
filter at this point - by hand. map is a bit crazy, as it turns out, so
it looks like the automatic incrementalizer thing is going to be
trickier than I've been letting myself imagine.