Fix calling convention. Add shuffle.
The calling convention I've been using for the Incrementalise
typeclasses is wrong. GHC does all the tyvars, then the typeclasses,
but I've been interspersing tyvars and typeclasses. I'm not doing
that any more (now it uses the GHC convention).

Added shuffle and mapWithKeys. This seems to work. It's promising as a
way of pushing the addressing problem (viz: which values are affected by
a given change) up to the programmer.